Se agrega persistencia DB a carrito

+++ 2daPreentrega/src/app.js
@@ -5,11 +5,16 @@ import { viewsRouter } from "../routes/viewsRoutes.js"
 import { Server as ServerIO } from "socket.io"
 import  __dirname from "../utils.js"
 import { create } from "express-handlebars"
-import { ProductManager } from "./ProductManager.js"
+import { ProductManager } from "../dao/ProductManager.js"
+import { persistencia } from "../utils.js"
+import { connectBD } from "../config/connectDB.js"
 import cors from "cors"
+import { productsModel } from "../dao/models/products.model.js"
+import { messagesModel } from "../dao/models/messages.model.js"
 
 const productManager= new ProductManager()
-const PORT = 8080
+const PORT = 8080 || window.location.port
+connectBD()
 const app = express()
 const handlebars = create({})
 
@@ -25,11 +30,12 @@ app.use('/api/carts/', cartRouter)
 app.use('/',viewsRouter)
 
 const httpServer= app.listen(PORT, ()=>{
-    console.log('Escuchando en el puerto 8080')
+    console.log(`Escuchando en el puerto ${PORT}`)
 })
 
 let productList=[]
 const io = new ServerIO(httpServer)
+let mensajes = []
 
 app.engine('handlebars', handlebars.engine)
 app.set('views',__dirname+'/views');
@@ -39,21 +45,65 @@ app.use(express.static(__dirname+'/public'))
 io.on('connection', socket=> {
     console.log('Nuevo cliente conectado');
     // Escuchando newProduct
-    socket.on('newProduct', data => {
+    socket.on('newProduct', async(data) => {
+        try
+        {if (persistencia=="FS"){
         let newProductId = productManager.addProduct(data)
         let newProduct = { ...data, id: newProductId }
-        productList.push(newProduct);
+        productList.push(newProduct)}
+
+        if (persistencia=="DB"){
+        await productsModel.create(data)
+        let id= await productsModel.findOne({title:data.title}).lean()._id
+        productList.push({...data,id:id})
+        }}catch(err){
+            console.log(err)
+        }
         
         // Emitiendo updateList
-        io.emit('updateList', productManager.getProducts())
+        io.emit('updateList', async()=>{
+            try{
+            if (persistencia=="FS"){
+            productManager.getProducts()}
+
+            if (persistencia=="DB"){
+            await productsModel.find({}).lean()
+            }}catch(err){
+                console.log(err)
+            }
+        })
         
-    });
+        
+    })
     // Escuchando deleteProducts
-    socket.on('deleteProduct', productId => {
-        const productList = productManager.getProducts();
-        const filteredList = productList.filter(product => product.id != productId.id);
-        productManager.deleteProduct(productId.id)
-        // Emitiendo updateList
-        io.emit('updateList', filteredList)
+    socket.on('deleteProduct', async (productId) => {
+        try{
+            if (persistencia=="FS"){
+            const productList = productManager.getProducts();
+            const filteredList = productList.filter(product => product.id != productId.id)
+            productManager.deleteProduct(productId.id)
+            // Emitiendo updateList
+            io.emit('updateList', filteredList)}
+
+            if (persistencia=="DB"){
+                const productList= await productsModel.find({}).lean()
+                const filteredList= productList.filter(product => product._id != productId._id)
+                await productsModel.findByIdAndDelete({_id:productId._id})
+                io.emit('updateList', filteredList)}
+            }
+        
+        catch(err)
+        {console.log(err)}
     });
+
+    socket.on('message', async(data) => {
+        try{
+        //mensajes.push(data)
+        await messagesModel.create(data)
+        mensajes=await messagesModel.find({}).lean()
+        io.emit('messageLogs', mensajes)}
+        catch(err){
+            console.log(err)
+        }
+    })
 })
